Prevent player from continuing if `currentMedia` was disposed
When `HTMLAudioElement.play()` is called the loaded media is transferred
to some buffer and the browser's internal player uses that buffer as
source. Merely cleaning up the resource for currentMedia (apparently)
isn't enough to stop the playback. So now, we clean up the resource,
transition to a hypothetical `stop` state where currentDuration and
currentTime is set to 0 and isPlaying is false. Then we set the src
attribute of HTMLAudioElement to an empty string (which, btw, isn't
cool because every next invocation of play will now spit out a nasty
error, but we can live with that for now).